fix(autoapi): keep table registry mutable
