+++ pkgs/standards/autoapi/autoapi/v3/__init__.py
@@ -74,7 +74,7 @@ from .system.dbschema import ensure_schemas, register_sqlite_attach, bootstrap_d
 from .config.constants import DEFAULT_HTTP_METHODS
 from .autoapi import AutoAPI
 
-from .orm.tables import Base
+from .table import Base
 from .types import App
