+++ pkgs/standards/autoapi/autoapi/v3/autoapi.py
@@ -102,6 +102,7 @@ class AutoAPI:
         self.rpc = SimpleNamespace()
         self.rest = SimpleNamespace()
         self.routers: Dict[str, Any] = {}
+        self.tables: Dict[str, Any] = {}
         self.columns: Dict[str, Tuple[str, ...]] = {}
         self.table_config: Dict[str, Dict[str, Any]] = {}
         self.core = SimpleNamespace()
@@ -341,8 +342,8 @@ class AutoAPI:
             md.create_all(bind=bind, checkfirst=True, tables=group)
 
         # 3) publish tables namespace
-        self.tables = SimpleNamespace(
-            **{
+        self.tables.update(
+            {
                 name: getattr(m, "__table__", None)
                 for name, m in self.models.items()
                 if hasattr(m, "__table__")
